+++ src/main/java/com/dungzi/backend/domain/review/dto/ReviewResponseDto.java
@@ -8,7 +8,7 @@ import lombok.Data;
 public class ReviewResponseDto {
     private String address;
 
-    private int totalRate;
+    private float totalRate;
 
     private String image;
